+++ src/main/java/ladder/domain/creator/RandomLadderCreator.java
@@ -1,14 +1,14 @@
-package ladder.creator;
+package ladder.domain.creator;
 
-import ladder.wrapper.NumberOfPerson;
-import ladder.wrapper.NumberOfRow;
-import ladder.wrapper.Row;
+import ladder.domain.wrapper.NumberOfPerson;
+import ladder.domain.wrapper.NumberOfRow;
+import ladder.domain.Row;
 
-public class SelfLadderCreator implements LadderCreator{
+public class RandomLadderCreator implements LadderCreator{
 
     private final Row[] rows;
 
-    public SelfLadderCreator(NumberOfRow row, NumberOfPerson numberOfPerson) {
+    public RandomLadderCreator(NumberOfRow row, NumberOfPerson numberOfPerson) {
         rows = new Row[row.getNumberOfRow()];
         for (int i = 0; i < row.getNumberOfRow(); i++) {
             rows[i] = new Row(numberOfPerson);
@@ -17,7 +17,7 @@ public class SelfLadderCreator implements LadderCreator{
 
     @Override
     public void drawLine(int x, int leftY, int rightY) {
-        rows[x].setValue(leftY, rightY);
+
     }
 
     @Override
