+++ src/main/java/ladder/domain/wrapper/LadderNumber.java
@@ -1,4 +1,4 @@
-package ladder.wrapper;
+package ladder.domain.wrapper;
 
 public class LadderNumber {
